azure/tests/requirements.txt
azure-data-tables==12.7.0
python-dotenv==1.2.1
playwright==1.58.0
pytest==9.0.2